WIP: Some refactors and implemented some actions and conditions.

+++ source/modules/game/module_game_manager.cpp
@@ -61,7 +61,7 @@ void CModuleGameManager::render()
 	// Replace this with separated menus
 	if (victoryMenuVisible) {
 
-		ImGui::SetNextWindowSize(ImVec2(window_width, window_height));
+		ImGui::SetNextWindowSize(ImVec2((float)window_width, (float)window_height));
 		ImGui::Begin("VICTORY!", false, window_flags);
 		ImGui::CaptureMouseFromApp(false);
 		ImGui::SetWindowPos("VICTORY!", ImVec2(menu_position.x, menu_position.y));
@@ -70,7 +70,7 @@ void CModuleGameManager::render()
 	}
 	else if (menuVisible) {
 
-		ImGui::SetNextWindowSize(ImVec2(window_width, window_height));
+		ImGui::SetNextWindowSize(ImVec2((float)window_width, (float)window_height));
 		ImGui::Begin("MENU", false, window_flags);
 		ImGui::CaptureMouseFromApp(false);
 		ImGui::SetWindowPos("MENU", ImVec2(menu_position.x, menu_position.y));
